refactor(Dialog-page): Move Page-component into the pages-folder
